server/requirements.txt
fastapi[standard]
python-dotenv
openai
pydantic